fix distance default, move rescue group routes to src directory, sort by animalupdated date
