+++ api_projects/templates/animals.html
@@ -31,19 +31,10 @@
                 <img src="{{ url_for('static', filename='imgs/shopping.png') }}" class="cart" alt="Cart">
             </a>
             </p>
-            {% if location %}
-                <p style="text-indent: 40px">Searching for cats in {{ location }}</p>
-                {% if distance %}
-                    <p style="text-indent: 40px">Searching for [Gender: {{ gender }}, Age: {{ age }}] cats in {{ location }} with radius of {{ distance }} miles</p>
-                {% else %}
-                    <p style="text-indent: 40px">Searching for [Gender: {{ gender }}, Age: {{ age }}] cats in {{ location }} with radius of 0 miles</p>
-                {% endif %}
+            {% if distance %}
+                <p style="text-indent: 40px">Searching for [Gender: {{ gender }}, Age: {{ age }}] cats in {{ location }} with radius of {{ distance }} miles</p>
             {% else %}
-                {% if distance %}
-                    <p style="text-indent: 40px">Searching for [Gender: {{ gender }}, Age: {{ age }}] cats in 48105 with radius of {{ distance }} miles</p>
-                {% else %}
-                    <p style="text-indent: 40px">Searching for [Gender: {{ gender }}, Age: {{ age }}] cats in 48105 with radius of 0 miles</p>
-                {% endif %}
+                <p style="text-indent: 40px">Searching for [Gender: {{ gender }}, Age: {{ age }}] cats in {{ location }} with radius of 0 miles</p>
             {% endif %}
         </div>
     </header>
