+++ api_projects/templates/error.html
@@ -9,5 +9,8 @@
         There was an error fetching your API Data
     </h2>
     <p>{{ error }}</p>
+    <p>Navigate back using link:</p>
+    <br>
+    <p>https://rescue-groups.herokuapp.com/animals/</p>
 </body>
 </html>
