--- common/utils/test.py
@@ -1,10 +1,0 @@
-from unittest import TestCase
-
-
-class Tests(TestCase):
-    """Test cases"""
-
-    def test(self):
-        """Test tests"""
-        test = True
-        self.assertEqual(test, True)
